requirements.txt
appdirs==1.4.3
appnope==0.1.0
click==6.7
decorator==4.0.11
Flask==0.12.1
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
oauthlib==2.0.2
packaging==16.8
pbr==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
Pygments==2.2.0
PyHamcrest==1.9.0
pyparsing==2.2.0
pytest==3.0.7
requests==2.13.0
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
traitlets==4.3.2
tweepy==3.5.0
wcwidth==0.1.7
Werkzeug==0.12.1
